Update pexp for ddns and add some custom functions
rc_check wont work because it uses the pgrep -x flag, which requires an exact
match. This fails because the daemons are running with '-c <path to kea.conf>'.
Adding the flag to pexp gets removed before the pgrep is run.

Also add rc_reload using the keactrl reload function.